% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_indicator.R
\name{time_to_indicator}
\alias{time_to_indicator}
\title{Internal function, Creates dbnp-import ready time indicater code from time
\code(time)}
\usage{
time_to_indicator(x1)
}
\description{
Internal function, Creates dbnp-import ready time indicater code from time
\code(time)
}

