% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbnp_generate_wide_save_file.R
\name{dbnp_generate_wide_save_file}
\alias{dbnp_generate_wide_save_file}
\title{Generate file specifically formatted for import into dbNp Phenotype Database}
\usage{
dbnp_generate_wide_save_file(x, id_var, treat_var, visit_var, time_var,
  visit_units = "weeks", time_units = "minutes", include_vars = NULL,
  exclude_vars = NULL, file = "", sep = ",", dec = ".",
  rename_id = TRUE)
}
\description{
Generate file specifically formatted for import into dbNp Phenotype Database
}

